Routes refactor

+++ erudite/core/routes/equipment.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter
 import logging
 from bson.objectid import ObjectId
 
-from ..db.models import Equipment, db, mongo_to_dict, ErrorResponseModel, ResponseModel, Response, mongo_to_dict_no_id
+from ..db.models import Equipment, db, mongo_to_dict, ErrorResponseModel, ResponseModel, Response, check_ObjectId
 
 router = APIRouter()
 
@@ -39,22 +39,21 @@ async def find_equipment(equipment_id: str):
     """Достаем обьект equipment из бд"""
 
     # Проверка на правильность ObjectId
-    try:
-        id = ObjectId(equipment_id)
-    except:
-        message = "ObjectId is written in the wrong format"
-        logger.info(message)
-        return ErrorResponseModel(400, message)
+    id = check_ObjectId(room_id)
 
-    # Проверка на наличие правилно введенного ObjectId в БД
-    equipment = await equipment_collection.find_one({"_id": id})
-    if equipment:
-        logger.info(f"Equipment {equipment_id}: {equipment}")
-        return ResponseModel(200, mongo_to_dict(equipment), "Room returned successfully")
+    if id:
+        # Проверка на наличие правилно введенного ObjectId в БД
+        equipment = await equipment_collection.find_one({"_id": id})
+        if equipment:
+            logger.info(f"Equipment {equipment_id}: {equipment}")
+            return ResponseModel(200, mongo_to_dict(equipment), "Room returned successfully")
+        else:
+            message = "This equipment is not found"
+            logger.info(message)
+            return ErrorResponseModel(404, message)
     else:
-        message = "This equipment is not found"
-        logger.info(message)
-        return ErrorResponseModel(404, message)
+        message = "ObjectId is written in the wrong format"
+        return ErrorResponseModel(400, message)
 
 
 @router.post(
@@ -89,22 +88,21 @@ async def delete_equipment(equipment_id: str):
     """Удаляем обьект equipment из бд"""
 
     # Проверка на правильность ObjectId
-    try:
-        id = ObjectId(equipment_id)
-    except:
-        message = "ObjectId is written in the wrong format"
-        logger.info(message)
-        return ErrorResponseModel(400, message)
+    id = check_ObjectId(room_id)
 
-    if await equipment_collection.find_one({"_id": id}):
-        await equipment_collection.delete_one({"_id": id})
-        message = f"Equipment: {equipment_id}  -  deleted from the database"
-        logger.info(message)
-        return ResponseModel(200, message, "Equipment deleted successfully")
+    if id:
+        if await equipment_collection.find_one({"_id": id}):
+            await equipment_collection.delete_one({"_id": id})
+            message = f"Equipment: {equipment_id}  -  deleted from the database"
+            logger.info(message)
+            return ResponseModel(200, message, "Equipment deleted successfully")
+        else:
+            message = f"Equipment: {equipment_id}  -  not found in the database"
+            logger.info(message)
+            return ErrorResponseModel(404, message)
     else:
-        message = f"Equipment: {equipment_id}  -  not found in the database"
-        logger.info(message)
-        return ErrorResponseModel(404, message)
+        message = "ObjectId is written in the wrong format"
+        return ErrorResponseModel(400, message)
 
 
 @router.patch(
@@ -118,22 +116,23 @@ async def patch_equipment(equipment_id: str, new_values: dict) -> str:
     """Обновляем необязательные поля в equipment в бд"""
 
     # Проверка на правильность ObjectId
-    try:
-        id = ObjectId(equipment_id)
-    except:
-        message = "ObjectId is written in the wrong format"
-        logger.info(message)
-        return ErrorResponseModel(400, message)
+    id = check_ObjectId(room_id)
 
-    if await equipment_collection.find_one({"_id": id}):
-        await equipment_collection.update_one({"_id": id}, {"$set": {"additional": new_values}})
-        message = f"Equipment: {equipment_id}  -  pached"
-        logger.info(message)  # Если ключа нет в обьекте, то будет добавлена новая пара ключ-значение к этому обьекту
-        return ResponseModel(200, message, "Equipment patched successfully")
+    if id:
+        if await equipment_collection.find_one({"_id": id}):
+            await equipment_collection.update_one({"_id": id}, {"$set": {"additional": new_values}})
+            message = f"Equipment: {equipment_id}  -  pached"
+            logger.info(
+                message
+            )  # Если ключа нет в обьекте, то будет добавлена новая пара ключ-значение к этому обьекту
+            return ResponseModel(200, message, "Equipment patched successfully")
+        else:
+            message = f"Equipment: {equipment_id}  -  not found in the database"
+            logger.info(message)
+            return ErrorResponseModel(404, message)
     else:
-        message = f"Equipment: {equipment_id}  -  not found in the database"
-        logger.info(message)
-        return ErrorResponseModel(404, message)
+        message = "ObjectId is written in the wrong format"
+        return ErrorResponseModel(400, message)
 
 
 @router.put(
@@ -147,30 +146,31 @@ async def update_equipment(equipment_id: str, new_values: Equipment):
     """Обновляем/добавляем поле/поля в equipment в бд"""
 
     # Проверка на правильность ObjectId
-    try:
-        id = ObjectId(equipment_id)
-    except:
-        message = "ObjectId is written in the wrong format"
-        logger.info(message)
-        return ErrorResponseModel(400, message)
+    id = check_ObjectId(room_id)
 
-    if await equipment_collection.find_one({"_id": id}):
-        await equipment_collection.delete_one({"_id": id})
-        await equipment_collection.insert_one({"_id": id})
-        await equipment_collection.update_one(
-            {"_id": id},
-            {
-                "$set": {
-                    "name": new_values.name,
-                    "type": new_values.type,
-                    "additional": mongo_to_dict_no_id(new_values.additional),
-                }
-            },
-        )
-        message = f"Equipment: {equipment_id}  -  updated"
-        logger.info(message)  # Если ключа нет в обьекте, то будет добавлена новая пара ключ-значение к этому обьекту
-        return ResponseModel(200, message, "Equipment updated successfully")
+    if id:
+        if await equipment_collection.find_one({"_id": id}):
+            await equipment_collection.delete_one({"_id": id})
+            await equipment_collection.insert_one({"_id": id})
+            await equipment_collection.update_one(
+                {"_id": id},
+                {
+                    "$set": {
+                        "name": new_values.name,
+                        "type": new_values.type,
+                        "additional": mongo_to_dict(new_values.additional),
+                    }
+                },
+            )
+            message = f"Equipment: {equipment_id}  -  updated"
+            logger.info(
+                message
+            )  # Если ключа нет в обьекте, то будет добавлена новая пара ключ-значение к этому обьекту
+            return ResponseModel(200, message, "Equipment updated successfully")
+        else:
+            message = f"Equipment: {equipment_id}  -  not found in the database"
+            logger.info(message)
+            return ErrorResponseModel(404, message)
     else:
-        message = f"Equipment: {equipment_id}  -  not found in the database"
-        logger.info(message)
-        return ErrorResponseModel(404, message)
+        message = "ObjectId is written in the wrong format"
+        return ErrorResponseModel(400, message)
